% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/makesyn.R
\name{make.syn}
\alias{make.syn}
\title{Construct synonym matrix for internal or external use}
\usage{
make.syn(return = TRUE)
}
\arguments{
\item{return}{Return value (T) or stout (F). Defaults to T.}
}
\description{
Construct synonym matrix for internal or external use
}

