% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{match}
\alias{match}
\title{Match terms}
\usage{
match(term_vector, pdf_tdm, term_tdm, thresholds, pdf_index, terms, sentences)
}
\arguments{
\item{term_vector}{Index vector of where each of the terms is in the pdf_tdm. i.e. the ith element of term_vector is j. Therefor, term i is at column j in the pdf_tdm.}

\item{pdf_tdm}{Term document matrix of words in the PDF}

\item{term_tdm}{Term document matrix of words in the terms and pdf sentences.}

\item{thresholds}{Acceptance thresholds}

\item{pdf_index}{Index of terms in PDF}

\item{terms}{List of terms used, this is the vector of column names of term_tdm.}

\item{sentences}{Vector of sentences read in from the PDf.}
}
\value{
List of matched terms.
}
\description{
Match terms in C++
}
