% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MF_terms.R
\docType{data}
\name{MF_terms}
\alias{MF_terms}
\title{Molecular Function terms
A dataset of molecular function terms.}
\format{An object of class \code{data.table} (inherits from \code{data.frame}) with 10834 rows and 1 columns.}
\usage{
MF_terms
}
\description{
Molecular Function terms
A dataset of molecular function terms.
}
\keyword{datasets}
