% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MF_terms_acession.R
\docType{data}
\name{MF_terms_accession}
\alias{MF_terms_accession}
\title{Molecular Function Terms with Accession numbers
Similar to the other data base but with assession numbers.}
\format{An object of class \code{data.table} (inherits from \code{data.frame}) with 10833 rows and 2 columns.}
\usage{
MF_terms_accession
}
\description{
Molecular Function Terms with Accession numbers
Similar to the other data base but with assession numbers.
}
\keyword{datasets}

