% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mineR_package.R
\docType{package}
\name{mineR_package}
\alias{mineR_package}
\alias{mineR_package-package}
\title{mineR: An R Package for Fuzzy Keyword Identification and Quantification in Natural Language}
\description{
Gene Ontology (GO) is a popular and frequently accessed database cataloging gene functions and processes. Groups across the world have used the data archived in GO to further understand and study the mechanisms behind life.  However, the annotation of GO terms and associations in the biomedical literature has been a persistent bottleneck for researchers; there is a growing need for easily accessible software packages which facilitate either fully or semi automated identification of GO terms present in the literature. Here we present an open source R package to address the above problem.  mineR, source found at  www.github.com/Chris1221/mineR, is programmed in R and C++ for performance.  This package addresses the persistent need of researchers to efficiently identify gene ontology terms in the literature, and can serve as a cog inside more complex workflows.
}
\section{Overview}{

mineR is a tool for identifying key terms in text. It has been developed with the intention of identifying ontological labels in free form text with specific application to finding Gene Ontology terms in the biomedical literature with strict canonical NLP quality control.
}

\section{Association}{

mineR associated terms with sentences through syntactical similarity. This means the number of words similar between the term and the sentence.
}

