% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replaceExpressions.character.R
\name{replaceExpressions.character}
\alias{replaceExpressions.character}
\title{Internal expression replacement with grep}
\usage{
\method{replaceExpressions}{character}(x)
}
\arguments{
\item{x}{pattern}
}
\description{
Internal expression replacement with grep
}
