% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replaceExpressions.PlainTextDocument.R
\name{replaceExpressions.PlainTextDocument}
\alias{replaceExpressions.PlainTextDocument}
\title{Internal expression replacement with grep}
\usage{
\method{replaceExpressions}{PlainTextDocument}(x)
}
\arguments{
\item{x}{pattern}
}
\description{
Internal expression replacement with grep
}
