% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replaceExpressions.R
\name{replaceExpressions}
\alias{replaceExpressions}
\title{Internal expression replacement with grep}
\usage{
replaceExpressions(x)
}
\arguments{
\item{x}{pattern}
}
\description{
Internal expression replacement with grep
}
