% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TDM.go.df.R
\docType{data}
\name{TDM.go.df}
\alias{TDM.go.df}
\title{Term document matrix of Gene Ontology Molecular Function terms}
\format{An object of class \code{data.frame} with 4987 rows and 10824 columns.}
\usage{
TDM.go.df
}
\description{
This is a provided data set for speeding up computations and also for use in unit testing. It has been compiled through the scripts available in \code{raw-data}
}
\keyword{datasets}
