% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makelim.R
\name{make.lim}
\alias{make.lim}
\title{Construct Constraint Limits for Fuzzy Identification}
\usage{
make.lim(int = TRUE, list = NULL, length = 10)
}
\arguments{
\item{int}{interactive (TRUE) or specified list (FALSE)}

\item{list}{preconstructed list if i is FALSE}

\item{length}{number of parameters you wish to enter. Defaults to 5.}
}
\description{
Construct Constraint Limits for Fuzzy Identification
}
