% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mineR.R
\name{mineR}
\alias{mineR}
\title{Text Miner}
\usage{
mineR(doc = character(), terms = character(), local = FALSE,
  lims = "interactive", output = character(), length = 10, wd = getwd())
}
\arguments{
\item{doc}{document which will be mined for keywords as a file path (character()).}

\item{terms}{list of \n seperated key phrases of varying lengths to be identified.}

\item{local}{currently only F is supported.}

\item{lims}{parameters for acceptance of identification (see FILL_IN)}

\item{output}{path to output file}

\item{length}{maximum length of term in words that you would like to search for.}

\item{wd}{working directory if different than getwd()}
}
\value{
A text document at wd/output if local == FALSE, or an R object if local == TRUE.
}
\description{
Fuzzy identification of key phrases from natural language
}
\examples{
demo(mineR)
}

