% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mineR.R
\name{mineR}
\alias{mineR}
\title{Text Miner}
\usage{
mineR(doc = NULL, terms = NULL, local = FALSE, lims = "interactive",
  output = NULL, syn = FALSE, syn.list = NULL, length = 10,
  wd = getwd(), return.as.list = FALSE, log = NULL, pdf_read = "R")
}
\arguments{
\item{doc}{document which will be mined for keywords as a file path (character()).}

\item{terms}{list of newline seperated key phrases of varying lengths to be identified.}

\item{local}{currently only F is supported.}

\item{lims}{Number of identical (or synonymous) words which must be present in a sentence in order for it to be accepted. "interactive" is default and allows you to interavtively build your own list, but a list of n elements can be supplied where n is the largest term you wish to search for.}

\item{output}{path to output file}

\item{syn}{TRUE if synonyms are to be used (also see syn.list), FALSE if no synonyms are to be used.}

\item{syn.list}{LIST of synonyms to be used. First element of each list item is the word that will counted if any of the other elements of that list item are present.}

\item{length}{maximum length of term in words that you would like to search for.}

\item{wd}{working directory if different than getwd()}

\item{return.as.list}{Mostly used for testing, returns output as an R (character) object instead of writing to a file.}
}
\value{
A text document at wd/output if local == FALSE, or an R object if local == TRUE.
}
\description{
Fuzzy identification of key phrases from natural language
}

